@@ -1,18 +1,12 @@
-import Package from '../package.json'
 import GlobalLiquid from './global/global'
 
 const MatterLiquid = {
-  name: Package.name,
-  version: Package.version,
+  name: PKG_NAME,
+  version: PKG_VERSION,
   for: 'matter-js@0.17.1',
-  // uses: [],
-  // options: {
-  //   something: true,
-  // },
   install(matter: any) {
     matter.Liquid = GlobalLiquid
   },
-}
+} as Matter.Plugin
 
-// @ts-expect-error Hard type flow
 Matter.Plugin.register(MatterLiquid)
